Fix byte-to-float conversion issue

+++ sensirion_common.h
@@ -93,15 +93,15 @@ extern "C" {
 uint32_t sensirion_bytes_to_uint32_t(const uint8_t* bytes);
 
 /**
- * sensirion_bytes_to_float() - Convert an array of bytes to an uint32_t
+ * sensirion_bytes_to_float() - Convert an array of bytes to a float
  *
  * Convert an array of bytes received from the sensor in big-endian/MSB-first
- * format to an uint32_t value in the correct system-endianness.
+ * format to an float value in the correct system-endianness.
  *
  * @param bytes An array of at least four bytes (MSB first)
- * @return      The byte array represented as uint32_t
+ * @return      The byte array represented as float
  */
-uint32_t sensirion_bytes_to_float(const uint8_t* bytes);
+float sensirion_bytes_to_float(const uint8_t* bytes);
 
 uint8_t sensirion_common_generate_crc(const uint8_t* data, uint16_t count);
